ViewportGadget : Simplify signature of gadgetsAt
Switching to Gadget* instead of GadgetPtr naturally cascades into the things that interact with gadgetsAt - I've also changed the functions which are private where it seems to make sense.  Definitely review this carefully, since it would be weird to introduce a crash in a rare corner case by switching the wrong thing, but I think all these make sense.