ViewportGadget : Bind new gadgetsAt signature
It would be consistent to bind this form as well, but it's not used anywhere, so maybe this commit should be omitted?  I would feel better about leaving it in if it was used in a test, but it looks like we don't do any ViewportGadget tests that require actually doing renders?