Spreadsheet : Add `spreadsheet:columnsNeedSerialisation` metadata
This can be used to avoid the serialisation of `setColumn()` calls on RowsPlugs, allowing them to be populated from a custom node's constructor instead.

This solution is not ideal, and the long term plan is to instead generalise the Serialiser with a generic mechanism for GraphComponents to define the serialisation requirements for any descendant individually (also allowing us to remove the `Plug::Dynamic` flag). There's enough value in the workaround that it's worth using for now though, rather than forcing people into workarounds of their own (see https://github.com/themissingcow/gaffer-astro/blob/master/src/GafferAstroModule/SpreadsheetSerialisation.cpp#L81).