GafferSceneUI::ContextAlgo::expandDescendants : Don't modify context in-place
This was the only place in Gaffer where an in-place modification of a context variable was done, and it added complexity, and a bug ( we were not signalling that the context had changed, because at the point when we called context->set(), the value matched the value we were setting it to, and so omitting the signal was skipped ).

I did a bunch of performance testing, and was unable to detect any scenario where there was a noticeable difference for this simpler code.