Widget : Add keyboard hotkey for starting and stopping drags
The extra calls to `__ensureMouseTracking()` and `__ensureEventFilter()` are necessary for features such as dragging onto a tab to switch tabs and continue the drag within. It seems that Qt gives us mouse move events on the source widget provided a mouse button is down, but if no buttons are down we need to get the mouse move events on the destination widget instead.