PathModel : Improve QVariant comparison
It wasn't at all clear that we were meeting the strict-weak-ordering requirements of `std::sort` when comparing mismatched types. We now compare by type first and then only compare values when the types match.

I ended up looking at this after getting crashes while working on other aspects of the sorting. In the end it wasn't the culprit because in practice we compare like types. But tightening this up and documenting the reason for avoiding the built-in variant comparison seems worthwhile anyway. Moving the comparison into its own function will also simplify future refactoring.