PathListingWidgetTest : Use QAbstractItemModelTester
This constantly monitors the model for consistency and turns any problems into test failures. This revealed two bugs that have gone unnoticed forever and which are fixed here :

- We were returning a valid index for `model->parent( childOfRoot )`. This was incorrect : the root must be represented by a default constructed `QModelIndex`, and we should never have an index where `index.internalPointer() == m_rootItem`.
- `rowCount()` was reporting children for all columns, instead of just the first one.

Although no logic is changed in `index()`, I've updated that to refer to a `parentItem` for consistency with the two functions fixed above.