WindowTest : Increase size of parent window
This avoids the following Qt errors on Mac :

```
ERROR : Qt : Window position QRect(93,-12 640x480) outside any known screen, using primary screen
```

It appears that the tiny parent window we were using before was so small that Qt couldn't figure out a sensible place to put the child. This isn't really a problem in real-world usage because we don't have windows so small.