ImageView : Add shortcuts for previous and next layer
This necessitated adjusting RGBAChannelsPlugValueWidget for subclassing, exposing the ordered list of labelled options, and allowing the layer menu to be extended. This also incidentally fixes a small bug : previously the checkbox was missing for the current layer when the menu had omitted the `RGBA` suffix (to avoid an unnecessary submenu).