PathModel : Evaluate Path asynchronously
We can only make changes to the model on the UI thread, so we move all evaluation of the path into a background task, and use queued events to make any necessary edits to the model. There are a few tricky parts to this due to dependencies on Item member data in `updateWalk()` itself :

- `Item::m_dataState` and `Item::m_childItemsState` : It is sufficient for these to use `std::atomic` because the only usage on the UI thread is to transition from Unrequested to Requested, and `compare_exchange_strong()` facilitates that.
- `Item::m_childItems` : We need to walk over the _new_ items, but they won't be applied to `m_childItems` until later. So we return the new items from `updateChildItems()`. This means holding the children by `shared_ptr`.
- `Item::m_displayData` : This is needed for sorting in `updateChildItems()`, but the queued updates to `m_displayData` are not visible immediately. So we return the appropriate value for sorting from `updateData()`.
- `Item::m_path` : Again, changes may not be applied until _after_ we need to use the path. It was always wasteful to store a Path instance for every single item anyway, so we now just store the item's name instead, and provide a path via `updateWalk()`. This does mean that we no longer support Paths which return a different subclass of Path in their `children()` implementation. But as documented on the Path class, that has never been properly supported.

The addition of `PathModel::waitForPendingUpdates()` also allows us to remove some of our reliance on the PathListingWidget for testing. Although we still need a widget to create the model, we no longer need to make it visible or rely on QTreeView to make queries to populate the model. The only model queries made are now our own, and we can wait reliably until all updates are done, instead of assuming that `waitForIdle()` is sufficient.

PathListingWidgetTest methods related to expansion and selection are expected failures at this point, because the expansion/selection code hasn't been adapted to work with the asynchronous model. That will happen in subsequent commits.