StandardStyle : Don't draw empty annotations
Empty bounds were causing us to draw huge frames with inverted min/max. We now simply skip the drawing of annotations with empty text. I did consider not returning them from `MetadataAlgo::annotations()` at all, but have held off on that for now - it means making the `annotations()` query slightly more expensive, and I'm not sure if someone might be expecting the current behaviour or not.