RendererAlgo : Support `IECoreScene::CoordinateSystem` in `objectSamples()`
This is necessary for the drawing of coordinate systems in the Viewer. This got broken in 3e4a513431d4ef85f4ef58d766a05bf77c161dda when adding motion blur support to IPR - the RenderController now calls `objectSamples()` rather than evaluating `objectPlug()->getValue()` directly.

This fix does expose the renderer backends to an object type they won't have seen previously, but they should be robust to that. I have included basic coverage for this in `ArnoldRenderTest.testCoordinateSystem()`. There is probably an argument for having `objectSamples()` be willing to sample everything except `NullObject`, but I've gone for the more conservative approach of whitelisting just CoordinateSystem for now.