CI : Update host OS to Ubuntu 20.04
As described at https://github.com/actions/virtual-environments/issues/3287, support for Ubuntu 16.04 was dropped by GitHub on September 20th. Our CI workflows are now failing to start because they can't find a worker.

We could update to either 18.04 or 20.04, and I don't know enough to make a truly informed decision between the two. Plumping for 20.04 since if it works, it will give us the longest period before we have to switch up again.

In practice this shouldn't affect us much at all, because we only use the host OS to launch a Docker container with our preferrred CentOS7 build environment, and all the real work happens in there.