Path : Add `cancellationSubject()` method
This facilitates path queries within BackgroundTasks, for paths which read from node graphs. It does feel like an oddity to include Plug in what is otherwise quite an abstract API. But there are three Path subclasses already which need to implement it, which is a reasonable proportion of the existing subclasses.

Other approaches considered were :

- Adding a GraphPath abstract class derived from Path, with a pure virtual `cancellationSubject()` method. This would keep the Path class graph-agnostic, but would mean any cancellation-aware clients would need to downcast to see if they have a GraphPath or not. This might be a bit purer, but in practice it just makes things a bit harder.
- Add a separate registration mechanism for getting the subject from a path, possibly in PathListingWidget. This would keep the Path base class agnostic, but make it harder for people to discover the functionality they need to implement.