LRUCache : Add support for cancellation
It was possible to perform cancellation before to some extent, by smuggling a canceller via `GetterKey`, and that is what we were doing for ValuePlug's internal cache. We are promoting this to be a first-class concept in LRUCache instead for two reasons :

- LRUCache is used to cache expensive work, and expensive work often needs cancellation support. The API should facilitate this rather than make it awkward.
- By supporting cancellation as a first-class concept, we can check for cancellation not only in the GetterFunction, but also in the internal `acquire()` method. This allows a client waiting on a concurrent call to `get()` to cancel in the meantime, as demonstrated in the new test cases.

This commit touches many files to update the function signatures for GetterFunctions, but these are boilerplate changes only. No client code (apart from the tests) is using the new cancellation feature yet.

The existing performance test methods in LRUCacheTest do show a slight regression, but of no more than 2%. The new performance tests with a canceller don't appear to show any impact relative to the equivalents without a canceller.