config/ie/options : Add IE_STOMP_VERSION
This tries to build to be compatible with the version in the current IE env, regardless of what version the SConstruct says this is supposed to be.  Useful when Gaffer has increased major version in preparation for possible compatibility breaks, but in practice it's still usable with the old version, and IE tools are not yet built for the new Gaffer version.