PathModel : Fix recursive expansion after recursive collapse
This bug could be reproduced easily as follows :

- View the GafferBot in the HierarchyView.
- Shift + click to expand fully from the root.
- Shift + click to collapse everything back to the root.
- Shift + click once more to expand fully. The expansion would only go one level deep this time.

The problem was that `collapseDescendants()` wasn't updating `Item::m_expandedInTreeView` so the next time we tried to expand we assumed that the QTreeView didn't need updating. Rather than add more tracking in `collapseDescendants()`, we instead move all the tracking to `treeViewExpanded()` and `treeViewCollapsed()` so that we update our state appropriately no matter how the change is made.

Also changed the tracking of selection state to follow the same principle. Although I'm not aware of a bug in that, this approach seems far more robust. It should mean that we don't need to touch `m_expandedInTreeView` or `m_selectedInSelectionModel` again.