PathListingWidget : Avoid `QItemSelectionModel`
QItemSelectionModel is full of quadratic performance hazards, which have been exacerbated by the async PathModel changes, because we were making many small edits to the selection as we progressed through the update. I did try a batching approach to limit the number of edits we made, and although this helped, it still had poor performance for certain operations (removing many items in particular, even when reduced to a single batched operation).

We now draw the selection directly based on queries to the PathMatcher, which allows us to also draw ancestors of the selection in a partially highlighted state at low cost. And we do our own event handling to implement editing of the selection. The end result : significantly improved performance.