SceneReader : Handle null return from `SceneInterface::readAttribute()`
This deals with a bug whereby USDScene advertises attributes that it can't actually load. We'll need to update Cortex to fix this at the root, but in the meantime we can at least make the SceneReader robust (the expectation throughout Gaffer and in CompoundObject itself is that there will be no null members).