PathListingWidget : Improve backward compatibility
The deprecated `get[Selected|Expanded]Paths()` methods work by querying the model, so we need to wait for the model update to complete. We also need to make sure that the model update has _started_, so we remove the lazy update from `__pathChanged`. This latter change is OK because PathModel itself is now doing lazy updates internally.

This fixes a couple of UIEditor bugs :

- Newly added plugs weren't shown in the plug editor panel.
- Newly added presets weren't shown in teh presets editor panel.

It would also have fixed an AnimationEditor bug if we hadn't recently refactored that to avoid the deprecated API.

The deprecated APIs do provide some convenience over the official API, by weeding out selected/expanded paths that don't currently exist in the model.
So I haven't yet refactored UIEditor to avoid them - I'll wait for feedback from other users of the deprecated APIs before deciding how to approach that.