NodeBinding : Fix `isInstanceOf()` to allow wrapping of NameSwitch
The approach we had taken to optimising `isInstanceOf( SwitchTypeId )` assumed that Switch would never be wrapped for derivation in Python, and we had a `static_assert()` to protect against this case. But it didn't protect against the wrapping of NameSwitch for derivation in Python, and unfortunately there is  such an extension class in a facility-specific module in the wild. This caused `customNameSwitch->isInstanceOf( SwitchTypeId )` to fail, resulting in an exception being thrown when creating a `SwitchPlugAdder` for the UI.

The solution is to avoid assumptions related to wrapping. Since we know the types we want to optimise are implemented in C++, we can still avoid entering Python, but by calling the base class implementation rather than returning `false` directly.