_Stylesheet : Work around Qt menu item alignment bug
Previously (Qt 5.12), Qt simply drew checkboxes in the padding area, with no knock-on effects to the positioning of the text. But now (Qt 5.15) the text will be shunted right by the width of the checkbox, misaligning it with any non-checkable items. We fix this in two steps :

- Fix the width of `QMenu::indicator`, so that checkable items have the same width in both checked and unchecked states.
- Add a negative margin on the right of the indicator to negate the unwanted offset.

See here for the bug report :

- https://bugreports.qt.io/browse/QTBUG-90242

And here for the eventual fix :

- https://codereview.qt-project.org/c/qt/qtbase/+/330178

I suspect the fix has been backported to one of the 5.15.x long term support releases but those are only accessible to folks with a commercial license. The open source version stops at 5.15.2.

There is also some discussion of the bug in the Qt forum :

- https://forum.qt.io/topic/121856/qmenu-items-checkable-not-checkable-not-text-aligned

In that forum an alternative workaround is suggested : adding a transparent icon to non-checkable items. We have rejected this approach for the following reasons :

- Our API allows specific icons to be added to menu items. In this case we can't add a transparent icon of the size we want anyway.
- Our preferred fix is more localised, affecting only the stylesheet for checkable items. Using an icon would affect Menu.py, which we wish to be style-agnostic.