ImageViewUI : Avoid QFontMetrics deprecation warning
This appeared in Qt 5.15. The official replacement for the `width()` method is `horizontalAdvance()`, but that's not actually the same as the space taken to draw the text - see https://doc.qt.io/qt-5/qfontmetrics.html#horizontalAdvance-1. Using `boundingRect()` seems to more accurately reflect the original intent, which was to give ourselves enough space to draw a particular string.