SConstruct : Prefer `-I` over `-isystem` for Arnold headers
This makes them visible to the SCons scanner, so that SCons can see when we are building with different Arnold versions, and can therefore cache build results appropriately.

The downside of not using `-isystem` is that the headers aren't opted out of the `-Werror` arguments we build with. This exposes a bug in Arnold's own headers in 6.2, whereby `ai_dotass.h` contains calls to deprecated functions. We work around this by including exactly the `ai_*.h` headers we need, instead of using the catch-all `ai.h`.