Arnold tests : Disable problematic tests for Arnold 7 on Mac
These exposed a bug in Arnold 7's render session handling. This could be reproduced with the following simple script :

```
arnold.AiBegin()

while True :

	# Make a universe, do an interactive render.
	# If we don't do this, then all is well.

	if True :

		universe = arnold.AiUniverse()
		arnold.AiSceneLoad( universe, './render.ass', None )
		session = arnold.AiRenderSession( universe )

		arnold.AiRenderBegin( session, arnold.AI_RENDER_MODE_CAMERA )
		arnold.AiRenderEnd( session )

		arnold.AiRenderSessionDestroy( session )
		arnold.AiUniverseDestroy( universe )

	# Make another universe, and load an ass file.
	# Occasionally this will fail, either crashing or
	# not producing any nodes.

	universe = arnold.AiUniverse()
	arnold.AiSceneLoad( universe, './load.ass', None )
	assert( arnold.AiNodeLookUpByName( universe, "test" ) is not None )
	arnold.AiUniverseDestroy( universe )

arnold.AiEnd()
```

We've confirmed that the upcoming Arnold 7.0.0.1 fixes both this simple repro and the original problem in the unit tests. But while that is still pending release, we want to be able to release Gaffer 0.61. Hence we disable the tests for the specific problematic Arnold version.