BranchCreator : Fix unnecessary input bounds evaluation
We were doing a complete (and completely unnecessary) recursive bounds evaluation of `/GAFFERBOT` when creating a branch at `/children`. In this case, `branchesData->locationOrAncestor( "/GAFFERBOT" )` correctly returned the root location, but we incorrectly interpreted that as meaning that `/GAFFERBOT` was an ancestor of the branch. If `path` is deeper than `location`, then path _cannot_ be the ancestor of a branch.