CatalogueUI : Derive Column from PathColumn
The idea behind PathColumn is that it provides a layer for translating the properties of a path into a form suitable for display in the UI as values, colours, tooltips etc. Because it didn't exist when CatalogueUI was written, we dealt with this back-to-front, by creating a "Column" class that actually added properties to the `_ImagesPath`.

Deriving `Column` from `PathColumn` has the following benefits :

- Catalogue columns can define background colours and tooltips.
- The CatalogueUI is consistent with the main PathColumn API.
- The implementation of `_ImagesPath` and `_ImageListing` is slightly simpler.

Backwards compatibility is retained by falling back to calling `self.value()` if `self._imageCellValue()` hasn't been implemented.