Documentation generators : Remove bogus targets
These are not actually generated by the scripts that claimed to generate them. This was forcing SCons to rerun the scripts for every build in the hope that the targets would somehow appear, which they didn't. And that in turn made `scons docs` unusable when trying to work on the copy.