Instancer : Remove unnecessary error in RootsPerVertex mode
We were erroring if the `prototypeRoots` variable was empty, but we don't need any prototypes if there are no vertices.

Removed the old test not only because it asserts the opposite of what we want, but because it was creating an invalid primitive variable in order to do so.

Fixes #4525.