SignalBinding : Avoid `slot_result_type`
The `slot_result_type` and the true `Result` type the signal was declared with are identical except when `Result` is `void`. In this case, Boost declares `slot_result_type` as `boost::detail::signals::unusable`, so that it can uniformly `return <something>;` rather than having to specialise to a plain `return;` for the void type.

We were using `slot_result_type` in our Python Slot and SlotCaller wrappers for the same reason, but C++17 and `if constexpr` give us a more elegant way of dealing with this without extra types and traits classes. Making this change will allow us to clean up many custom slot callers, and also smooths the path towards replacing `boost::signals` entirely.