LightEditor : Fix section chooser bug
There were two bugs here. Firstly, we were doing the wrong mapping from the value of the `section` plug to the name of the section. This made it less likely that we'd see the second bug, where we tried to call a completely nonexistent function.

This did make me question the whole idea of allowing parameters to be registered with no section. If we are mapping "no section" to "Main" internally, why not just force people to register it that way? But I think we do want to stick with it. One part of the LightEditor roadmap is to introduce an alternate view more akin to the NodeEditor, with all parameters visible in collapsible sections. At that point we do need to distinguish between "no section" and an explicit "Main" subsection.