SpreadsheetUI : Improve useability of read-only Spreadsheets
Before we were simply disabling the entire widget, which made it impossible to switch between sections and scroll to see all the rows. We now keep the widget enabled and disable the specific editing actions within it.

Fixes #4529. Note that the issue also complains of read-only spreadsheets being very dark with hard-to-read text. This PR has improved the contrast slightly by virtue of not calling `setEnabled( False )`, but it's still quite subtle.