Signal : Deprecate default value for `scoped` argument
Instead, require that either `scoped = True` or `scoped = False` is passed explicitly. There's a bit of a long back story to this one...

When I first made the signal bindings for Python way back in the mists of time, I made the mistake of _not_ mirroring the C++ API directly. In C++, `connect()` returns a `Connection` object which can be used to manage the connection. In Python I returned a `ScopedConnection` instead, which is like `Connection` but _also_ disconnects the slot automatically when the `ScopedConnection` is destroyed. This naturally lead to confusion :

```
# The slot will never be called, because the returned ScopedConnection is
# destroyed immediately.
signal.connect( mySlot )
# The slot will be called as expected, but finding somewhere to store
# `connection` can be inconvenient, particularly in config files.
connection = signal.connect( mySlot )
```

In `0.14.0.0` we added a `scoped` argument, which can be used to request the same behaviour as the C++ default, with a regular `Connection` object being returned from `connect()`.

```
# Works!
signal.connect( mySlot, scoped = False )
# Still doesn't work - the default is `scoped = True`
signal.connect( mySlot )
```

In `0.54.0.0` we went further, and introduced a new Trackable Python class that negated one of the major reasons for wanting to use `scoped = True` in Python (#3179). By deriving from Trackable, classes get automatic disconnection of their slots on class destruction, regardless of what type of connection they made. This closely matches the typical C++ usage, which is what we are striving for in the Python bindings.

Despite these improvements, the original problem remains : the default behaviour of `connect()` is confusing and unhelpful, and continues to trip people up, new and old hands alike. We really want the default to be `scoped = False`. In the absence of a time machine, this PR makes the first step in this direction by deprecating the usage of the default value. Once everyone has updated to use `scoped = something` universally, we can then change the default value to `scoped = False` in a future version without fear of breaking everything.

The deprecation is handled by emitting DeprecationWarning via Python's standard `warnings` mechanism, in `src/GafferBindings/SignalBinding.cpp`. This warning is turned into a failure in `gaffer test`, and printed to the terminal in a regular `gaffer` session. Downstream users can configure alternative behaviours using `warnings.simplefilter()` etc.

After updating all call sites in Gaffer appropriately, we are left with 645 calls with `scoped = False` and 173 with `scoped = True`. This provides further weight to the idea that `False` is the correct default.