SlotCallIterator : Don't store `m_end`
Instead, use a sentinel `m_slot( nullptr )` to represent the end iterator for a range, and consider `!m_slot->next` to signify the end when advancing an iterator from the start of the range (this is the empty `Signal::m_lastSlot`, as all other slots have a valid `next`).

The reduction in reference counting operations yields a further 3x speedup in `SignalsTest.testCallPerformance`.