SceneHistoryUI : Allow "Edit Source" to show referenced nodes
When we designed this, we thought the priority was to show a node that could truly be edited. So when the source node was inside a Reference, we showed the reference node for editing, on the assumption that some plugs of the source node would have been promoted for editing. But in practice, this is rarely the case, and when using "Edit Source" in this scenario, users want to see the true source node still so that they can inspect its parameters etc.