SlotBase : Use an atomic reference count
This is sufficient to allow concurrent emission of the same signal from multiple threads, which is necessary to provide temporary support for some legacy code.

This does have a significant impact on `testCallPerformance` relative to
our non-atomic version, but still yields a 40% improvement over
`boost::signals` in that benchmark. `testConnectionPerformance` sees a slight drop, and `testConstructionPerformance` and `testMany` show no difference.