CI : Upload artifacts even for release builds
We publish build tarballs in two places :

- As artifacts on the PR itself, so people can test the PR functionality.
- On the GitHub Releases page for official releases. This only happens for builds corresponding to tagged releases.

The publishing for the latter has been unreliable lately, sometimes timing out before it is complete. But we will now at least have the PR artifact as a backup, allowing us to do a manual upload to the release if necessary.

I have yet to determine why artifact uploads are reliable and release uploads are not. Googling suggests we're not completely alone, and that publishing from within a Docker container might be part of the problem. But I haven't been able to get to the bottom of it yet.