NodeSetEditor : Always show node names in tab label
This essentially reverts 7941677175a07fb7ae2248b651b66f00091bee74, where we removed them because :

- The tabs were taking up a lot of space, and felt a bit cluttered. Particularly in the NodeEditor/SceneInspector panel, there often wasn't enough space to show them all.
- The tab sizes changing every time you choose a node is distracting.

Bringing the names back as an experiment in response to a request from a user. I have some sympathy for this as I myself have sometimes missed the "at a glance" confirmation of what is in the Viewer. But having tried it, I can't say I like it - I prefer the cleaner look, and I don't think we should merge. One justification for why it might be more acceptable this time around is we now have the burger menu to choose tabs when there isn't enough space to show them all. But that still doesn't swing it for me.

Other possibilities :

- Keep the names, but use a smaller/fainter font for them. This is tricky because QTabBar doesn't support rich text out of the box. I've had some success inserting a QLabel instead, but am currently thwarted by layout/sizing issues. Even if this works, it will still take space, and still jump in size when swapping nodes.
- Make sure all editors show the node name internally somewhere, so it's not needed in the tab label. This is already the case for the NodeEditor,SceneInspector and PrimitiveInspector.
	- One logical extension of this might be to move the editor focus menu out of the tab bar, so that each editor has their own focus menu internally. Originally it was in the tab bar to save space, but it would be more logical inside the editor, since it modifies properties of the editor.
- Display the name of the focus node globally somewhere, perhaps in the top right of the main menu? This isn't what the user requested exactly, but it does seem that their motivation was to know what was focussed, when jumping focus between bookmarks frequently.
	- This seems related to the problem of where to choose an EditScope. We've already established that choosing a scope in each editor is going to become confusing as we make more editors scope-aware. I quite like the idea of a combined focus/editscope widget in the top right of the main menu. I also wonder if editscopes should be tied to the focus somehow, so switching focus automatically switches scope.

All this is to say, "I have tried, here's a PR if you'd like to see it, but on balance I don't think we should merge it". But I do think there are a bunch of interrelated questions here that we will have to address one way or another soonish.