ArnoldShader : Add support for loading imagers
While strictly speaking imagers are not shaders, they have parameters and outputs and they can be connected together to form chains, so the natural representation for them in the Gaffer datastream is a ShaderNetwork. Since ArnoldShader is designed to show parameters and outputs and generate ShaderNetworks from them, it seems reasonable to reuse it for imagers.