Arnold Renderer : Improve logging & messaging tests
Logging: We only needed to skip the profile.json portion of this test in Ai 7.0 and it appears to still be broken in 7.1 so I have relaxed the logic, lets check again in Ai 8.0

Messaging: We only need to skip the SceneDescription portion of this test. Unfortunately this is still not fixed in Ai 7.1 and there is no timeline on a fix.