VectorDataWidget : Fix display of mismatched vector lengths
This was truncating the number of rows to the length of the first column, and printing errors like this :

```
IndexError: Index out of range
Traceback (most recent call last):
  File "/home/john/dev/build/gafferPython3/python/GafferUI/VectorDataWidget.py", line 909, in data
    return column.accessor.getElement( index.row(), column.relativeColumnIndex )
  File "/home/john/dev/build/gafferPython3/python/GafferUI/VectorDataWidget.py", line 957, in getElement
    return GafferUI._Variant.toVariant( self.data()[rowIndex] )
```

Strictly speaking, we never intended to support mismatched lengths, but they can occur in at least two places :

- The PrimitiveInspector Constant section, where a mixture of scalar and vector variables might be shown, and the vector variables can be of arbitrary length.
- The `RandomChoice.choices` widget, if an expression has been used to drive `choices.values` but not `choices.weights`.

In both cases, the VectorDataWidget is read only, so I haven't implemented editing for mismatched length vectors. I think it could probably be achieved by using `_DataAccessor.defaultElement()` to resize the column when necessary, if in future we find a use case we can test it with.