GraphEditor : Hide focus gadget for auxiliary nodes
Because such nodes are small, it is too easy to accidentally hit the focus gadget instead when trying to select them. Rather than hide the gadget unconditionally, the visibility is controlled by metadata. This gives us more fine-grained control for future use, and a get-out-of-jail-free card if somebody prefers the old behaviour.