StyleSheet : Remove QHeaderView height hack
I don't know the history for this, but it's causing more trouble than its worth now. In Qt 5.12 it causes `QStyleSheetStyle::sizeFromContents()` to completely ignore the header contents (even in the horizontal direction), which in turn means that PathListingWidget headings are way too narrow, which means the LightEditor is basically unreadable. This isn't an issue in our standard Qt 5.15 builds where `sizeFromContents()` is fixed, but certain folks are partial to their custom retro Qt 5.12 builds.

Instead of hacking the height, we now just ask for zero padding in the vertical direction. While it sure looks like there's still a bit of padding, the sizes are reasonable and are a close enough match to the old behaviour, in both Qt 5.12 and 5.15.

This also exposed a small bug in the Spreadsheet UI - we need to specify a vertical alignment for the header, otherwise it gets aligned to the top.