SceneViewTest : Update workaround following signal changes
Now that exceptions in signal handling are output via `IECore.msg()`, our test message handler detects that this test has been troubled for a while. I did spend a little time trying to get to the bottom of the problem (something to do with mismatched GafferUI/Qt ownership in `GLWidget.addOverlay()` perhaps?) but in the end this wasn't fruitful. So instead I've extended the workaround with a new workaround (the new workaround is needed for Python 3, and the old for Python 2).