Container : Fix symbol exports
This was broken by the Windows fix in c16f7095d5df22e2cc36ed4c83fd6edf8165301f, but went unnoticed because we don't currently rely on any Container symbols outside of `libGaffer`. That changes in https://github.com/GafferHQ/gaffer/pull/4358 where we want to instantiate Container in a library other than `libGaffer` and then reference those symbols in a Python module.

We now treat Container in the same way as TypedPlug and TypedObjectPlug :

- Don't include `Container.inl` in any public headers, because that would lead to duplicate symbols.
- Instead include `Container.inl` in the `.cpp` file where we wish to make a particular instantiation.
- Apply IECORE_EXPORT to the Container class, so that all instantiations are exported, regardless of which module they are made in.