ArnoldLightVisualiser : Include Arnold headers so IES vis works
I'm not sure exactly when a dependency shift caused these headers to no longer be included, but unexpectedly, because all the code that uses the Arnold API is behind a check for AI_VERSION_ARCH_NUM >= 6, without the Arnold headers, rather than getting a compile error, it simply omits the IES visualiser.  Weird symptom, but easy fix.