PlugAlgo : Support NameValuePlug in `extractDataFromPlug()`
This fixes #4583, where `Spreadsheet.resolvedRows` was failing to compute if the Spreadsheet contained a NameValuePlug.

I did consider adding the new CompoundData fallback for all unknown plug types that had children, but decided against it. It's not 100% clear that such plugs wouldn't have a better dedicated conversion of their own. Adding the conversion we want later will be easier if we don't add an unwanted conversion now (it could become a compatibility hazard).