SConstruct : Perform header substitutions before installation
Without this, it was impossible to `#include "Gaffer/Version.h"` from any Gaffer source file, because the header was only substituted after installation.

This also allows us to remove the `-DGAFFER_VERSION=...` compiler arguments in favour of `Version.h`, which means less churn in build caches when version numbers change.